+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - IMPROVED
   - You can now disable auto init (Axentix method) with class .no-axentix-init, and manually init element
   - Sidenav link height increased to make it more SEO friendly
+  - Modal have now bodyScrolling option like Sidenav by default
 - BUGFIXES
   - Button's height broken in sidenav